Test

See merge request comprehensive/galaplat-comprehensive-bidding!125

+++ galaplat-comprehensive-bidding-service/src/main/java/com/galaplat/comprehensive/bidding/vos/JbxtBiddingVO.java
@@ -39,13 +39,13 @@ public class JbxtBiddingVO implements Serializable {
 
      private Integer isdelay;
 
-     private Integer bidPercent;
+     private BigDecimal bidPercent;
 
-     public Integer getBidPercent() {
+     public BigDecimal getBidPercent() {
          return bidPercent;
      }
 
-     public void setBidPercent(Integer bidPercent) {
+     public void setBidPercent(BigDecimal bidPercent) {
          this.bidPercent = bidPercent;
      }
